+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## UNRELEASED
 
+### ENHANCEMENTS
+
+* Support Alien4Cloud 3.2.0 ([GH-723](https://github.com/ystia/yorc/issues/723))
+
+
 ## 4.1.0 (April 11, 2021)
 
 ### DEPENDENCIES